+++ .vscode/settings.json
@@ -168,5 +168,5 @@
     "__tuple": "cpp",
     "__verbose_abort": "cpp"
   },
-  "cmake.sourceDirectory": "${workspaceFolder}/packages/realm_dart",
+  "cmake.sourceDirectory": "${workspaceFolder}/packages/realm_dart"
 }
