+++ packages/realm_dart/lib/src/configuration.dart
@@ -658,7 +658,7 @@ class ClientResetError extends SyncError {
       throw RealmException("Missing `originalFilePath`");
     }
 
-    return _app.handle.immediatelyRunFileActions(originalFilePath!);
+    return _app.handle.resetRealm(originalFilePath!);
   }
 }
