+++ packages/realm_dart/lib/src/native/app_handle.dart
@@ -2,22 +2,20 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:ffi';
 import 'dart:io';
 import 'dart:isolate';
 
-import 'ffi.dart';
-
 import '../init.dart';
-import '../logging.dart';
 import '../realm_class.dart';
 import '../scheduler.dart';
 import 'convert.dart';
 import 'convert_native.dart';
 import 'credentials_handle.dart';
 import 'error_handling.dart';
+import 'ffi.dart';
 import 'handle_base.dart';
+import 'http_transport_handle.dart';
 import 'realm_bindings.dart';
 import 'realm_core.dart';
 import 'realm_library.dart';
@@ -34,7 +32,7 @@ class AppHandle extends HandleBase<realm_app> {
       _firstTime = false;
       realmLib.realm_clear_cached_apps();
     }
-    final httpTransportHandle = _createHttpTransport(configuration.httpClient);
+    final httpTransportHandle = HttpTransportHandle.from(configuration.httpClient);
     final appConfigHandle = _createAppConfig(configuration, httpTransportHandle);
     return AppHandle(realmLib.realm_app_create_cached(appConfigHandle.pointer));
   }
@@ -88,7 +86,7 @@ class AppHandle extends HandleBase<realm_app> {
           createAsyncCallbackUserdata(completer),
           realmLib.addresses.realm_dart_userdata_async_free,
         )
-        .raiseLastErrorIfFalse("Remove user failed");
+        .raiseLastErrorIfFalse();
     return completer.future;
   }
 
@@ -99,7 +97,7 @@ class AppHandle extends HandleBase<realm_app> {
             pointer,
             user.pointer,
           )
-          .raiseLastErrorIfFalse("Switch user failed");
+          .raiseLastErrorIfFalse();
     });
   }
 
@@ -121,7 +119,7 @@ class AppHandle extends HandleBase<realm_app> {
             createAsyncCallbackUserdata(completer),
             realmLib.addresses.realm_dart_userdata_async_free,
           )
-          .raiseLastErrorIfFalse("Update base URL failed");
+          .raiseLastErrorIfFalse();
     });
     return completer.future;
   }
@@ -136,7 +134,7 @@ class AppHandle extends HandleBase<realm_app> {
           createAsyncCallbackUserdata(completer),
           realmLib.addresses.realm_dart_userdata_async_free,
         )
-        .raiseLastErrorIfFalse("Refresh custom data failed");
+        .raiseLastErrorIfFalse();
     return completer.future;
   }
 
@@ -144,7 +142,7 @@ class AppHandle extends HandleBase<realm_app> {
     return realmLib.realm_app_get_app_id(pointer).cast<Utf8>().toRealmDartString()!;
   }
 
-  Future<UserHandle> logIn(CredentialsHandle credentials) async {
+  Future<UserHandle> logIn(CredentialsHandle credentials) {
     final completer = Completer<UserHandle>();
     realmLib
         .realm_app_log_in_with_credentials(
@@ -154,8 +152,8 @@ class AppHandle extends HandleBase<realm_app> {
           createAsyncUserCallbackUserdata(completer),
           realmLib.addresses.realm_dart_userdata_async_free,
         )
-        .raiseLastErrorIfFalse("Login failed");
-    return await completer.future;
+        .raiseLastErrorIfFalse();
+    return completer.future;
   }
 
   Future<void> registerUser(String email, String password) {
@@ -175,7 +173,7 @@ class AppHandle extends HandleBase<realm_app> {
     return completer.future;
   }
 
-  Future<void> confirmUser(String token, String tokenId) async {
+  Future<void> confirmUser(String token, String tokenId) {
     final completer = Completer<void>();
     using((arena) {
       realmLib
@@ -189,7 +187,7 @@ class AppHandle extends HandleBase<realm_app> {
           )
           .raiseLastErrorIfFalse();
     });
-    return await completer.future;
+    return completer.future;
   }
 
   Future<void> resendConfirmation(String email) {
@@ -276,31 +274,6 @@ class AppHandle extends HandleBase<realm_app> {
     return completer.future;
   }
 
-  Future<void> logOut(UserHandle? user) {
-    final completer = Completer<void>();
-    if (user == null) {
-      realmLib
-          .realm_app_log_out_current_user(
-            pointer,
-            realmLib.addresses.realm_dart_void_completion_callback,
-            createAsyncCallbackUserdata(completer),
-            realmLib.addresses.realm_dart_userdata_async_free,
-          )
-          .raiseLastErrorIfFalse("Logout failed");
-    } else {
-      realmLib
-          .realm_app_log_out(
-            pointer,
-            user.pointer,
-            realmLib.addresses.realm_dart_void_completion_callback,
-            createAsyncCallbackUserdata(completer),
-            realmLib.addresses.realm_dart_userdata_async_free,
-          )
-          .raiseLastErrorIfFalse("Logout failed");
-    }
-    return completer.future;
-  }
-
   Future<void> deleteUser(UserHandle user) {
     final completer = Completer<void>();
     realmLib
@@ -311,11 +284,11 @@ class AppHandle extends HandleBase<realm_app> {
           createAsyncCallbackUserdata(completer),
           realmLib.addresses.realm_dart_userdata_async_free,
         )
-        .raiseLastErrorIfFalse("Delete user failed");
+        .raiseLastErrorIfFalse();
     return completer.future;
   }
 
-  bool immediatelyRunFileActions(String realmPath) {
+  bool resetRealm(String realmPath) {
     return using((arena) {
       final didRun = arena<Bool>();
       realmLib
@@ -324,7 +297,7 @@ class AppHandle extends HandleBase<realm_app> {
             realmPath.toCharPtr(arena),
             didRun,
           )
-          .raiseLastErrorIfFalse("An error occurred while resetting the Realm. Check if the file is in use: '$realmPath'");
+          .raiseLastErrorIfFalse();
       return didRun.value;
     });
   }
@@ -347,152 +320,6 @@ class AppHandle extends HandleBase<realm_app> {
       return completer.future;
     });
   }
-}
-
-// TODO:
-// We need a pure Dart equivalent of:
-// ```dart
-// ServiceBinding.rootIsolateToken != null
-// ```
-// to get rid of this hack.
-final bool _isRootIsolate = Isolate.current.debugName == 'main';
-
-class _HttpTransportHandle extends HandleBase<realm_http_transport> {
-  _HttpTransportHandle(Pointer<realm_http_transport> pointer) : super(pointer, 24);
-}
-
-_HttpTransportHandle _createHttpTransport(HttpClient httpClient) {
-  final requestCallback = Pointer.fromFunction<Void Function(Handle, realm_http_request, Pointer<Void>)>(_requestCallback);
-  final requestCallbackUserdata = realmLib.realm_dart_userdata_async_new(httpClient, requestCallback.cast(), scheduler.handle.pointer);
-  return _HttpTransportHandle(realmLib.realm_http_transport_new(
-    realmLib.addresses.realm_dart_http_request_callback,
-    requestCallbackUserdata.cast(),
-    realmLib.addresses.realm_dart_userdata_async_free,
-  ));
-}
-
-void _requestCallback(Object userData, realm_http_request request, Pointer<Void> requestContext) {
-  //
-  // The request struct only survives until end-of-call, even though
-  // we explicitly call realm_http_transport_complete_request to
-  // mark request as completed later.
-  //
-  // Therefore we need to copy everything out of request before returning.
-  // We cannot clone request on the native side with realm_clone,
-  // since realm_http_request does not inherit from WrapC.
-
-  final client = userData as HttpClient;
-
-  client.connectionTimeout = Duration(milliseconds: request.timeout_ms);
-
-  final url = Uri.parse(request.url.cast<Utf8>().toRealmDartString()!);
-
-  final body = request.body.cast<Utf8>().toRealmDartString(length: request.body_size);
-
-  final headers = <String, String>{};
-  for (int i = 0; i < request.num_headers; ++i) {
-    final header = request.headers[i];
-    final name = header.name.cast<Utf8>().toRealmDartString()!;
-    final value = header.value.cast<Utf8>().toRealmDartString()!;
-    headers[name] = value;
-  }
-
-  _requestCallbackAsync(client, request.method, url, body, headers, requestContext);
-  // The request struct dies here!
-}
-
-Future<void> _requestCallbackAsync(
-  HttpClient client,
-  int requestMethod,
-  Uri url,
-  String? body,
-  Map<String, String> headers,
-  Pointer<Void> requestContext,
-) async {
-  await using((arena) async {
-    final responsePointer = arena<realm_http_response>();
-    final responseRef = responsePointer.ref;
-    final method = HttpMethod.values[requestMethod];
-
-    try {
-      // Build request
-      late HttpClientRequest request;
-
-      switch (method) {
-        case HttpMethod.delete:
-          request = await client.deleteUrl(url);
-          break;
-        case HttpMethod.put:
-          request = await client.putUrl(url);
-          break;
-        case HttpMethod.patch:
-          request = await client.patchUrl(url);
-          break;
-        case HttpMethod.post:
-          request = await client.postUrl(url);
-          break;
-        case HttpMethod.get:
-          request = await client.getUrl(url);
-          break;
-      }
-
-      for (final header in headers.entries) {
-        request.headers.add(header.key, header.value);
-      }
-
-      if (body != null) {
-        request.add(utf8.encode(body));
-      }
-
-      Realm.logger.log(LogLevel.debug, "HTTP Transport: Executing ${method.name} $url");
-
-      final stopwatch = Stopwatch()..start();
-
-      // Do the call..
-      final response = await request.close();
-
-      stopwatch.stop();
-      Realm.logger.log(LogLevel.debug, "HTTP Transport: Executed ${method.name} $url: ${response.statusCode} in ${stopwatch.elapsedMilliseconds} ms");
-
-      final responseBody = await response.fold<List<int>>([], (acc, l) => acc..addAll(l)); // gather response
-
-      // Report back to core
-      responseRef.status_code = response.statusCode;
-      responseRef.body = responseBody.toCharPtr(arena);
-      responseRef.body_size = responseBody.length;
-
-      int headerCnt = 0;
-      response.headers.forEach((name, values) {
-        headerCnt += values.length;
-      });
-
-      responseRef.headers = arena<realm_http_header>(headerCnt);
-      responseRef.num_headers = headerCnt;
-
-      int index = 0;
-      response.headers.forEach((name, values) {
-        for (final value in values) {
-          final headerRef = (responseRef.headers + index).ref;
-          headerRef.name = name.toCharPtr(arena);
-          headerRef.value = value.toCharPtr(arena);
-          index++;
-        }
-      });
-
-      responseRef.custom_status_code = CustomErrorCode.noError.code;
-    } on SocketException catch (socketEx) {
-      Realm.logger.log(LogLevel.warn, "HTTP Transport: SocketException executing ${method.name} $url: $socketEx");
-      responseRef.custom_status_code = CustomErrorCode.timeout.code;
-    } on HttpException catch (httpEx) {
-      Realm.logger.log(LogLevel.warn, "HTTP Transport: HttpException executing ${method.name} $url: $httpEx");
-      responseRef.custom_status_code = CustomErrorCode.unknownHttp.code;
-    } catch (ex) {
-      Realm.logger.log(LogLevel.error, "HTTP Transport: Exception executing ${method.name} $url: $ex");
-      responseRef.custom_status_code = CustomErrorCode.unknown.code;
-    } finally {
-      realmLib.realm_http_transport_complete_request(requestContext, responsePointer);
-    }
-  });
 }
 
 Pointer<Void> createAsyncFunctionCallbackUserdata(Completer<String> completer) {
@@ -524,7 +351,7 @@ void _callAppFunctionCallback(Pointer<Void> userdata, Pointer<Char> response, Po
   completer.complete(stringResponse);
 }
 
-Pointer<Void> createAsyncCallbackUserdata<T extends Function>(Completer<void> completer) {
+Pointer<Void> createAsyncCallbackUserdata(Completer<void> completer) {
   final callback = Pointer.fromFunction<
       Void Function(
         Pointer<Void>,
@@ -555,7 +382,7 @@ class _AppConfigHandle extends HandleBase<realm_app_config> {
   _AppConfigHandle(Pointer<realm_app_config> pointer) : super(pointer, 8);
 }
 
-_AppConfigHandle _createAppConfig(AppConfiguration configuration, _HttpTransportHandle httpTransport) {
+_AppConfigHandle _createAppConfig(AppConfiguration configuration, HttpTransportHandle httpTransport) {
   return using((arena) {
     final appId = configuration.appId.toCharPtr(arena);
     final handle = _AppConfigHandle(realmLib.realm_app_config_new(appId, httpTransport.pointer));
@@ -590,3 +417,11 @@ _AppConfigHandle _createAppConfig(AppConfiguration configuration, _HttpTransport
     return handle;
   });
 }
+
+// TODO:
+// We need a pure Dart equivalent of:
+// ```dart
+// ServiceBinding.rootIsolateToken != null
+// ```
+// to get rid of this hack.
+final bool _isRootIsolate = Isolate.current.debugName == 'main';
